added support for a couple of benchmarking frameworks
Nonius and Celero are a couple of C++ frameworks for writing benchmarks,
similar to writing unit tests.

Nonius is a header-only library, so it is easy to include in
a project. It has a nice statistical analysis in which benchmarks are
automatically run multiple times and an estimate of the mean and
standard deviation of runtime is computed.

Celero requires compilation, so it has to be built with the project.
However, it has a really nice "Experiments" feature that allows you to
to and experiment comparing several different pieces of code to some
baseline. This makes it useful for quickly testing a couple of different
algorithms when developing code to decide which one to use.